A number of things
Added remove buttons to both watchlist and holding info (functionality was there just forgot to put the buttons). Added button and functionality for user to share watchlist, user is prompted to enter an email and after submitting the email is sent sharing the contents of the watchlist. Performance of stocks on the user watchlist is now charted with a bar chart showing the % change from when the stock was added to the current price