Many changes
Noticed static files were in the .gitignore, not ideal so those are being committed. Separated CSS from html files to their own files, added logout functionality which had a URL but no actual view. Added a total percent change to display on dashboard. The static files in the gitignore likely caused the issues with the dashboard display